chore(shadertools): Remove ShaderModuleInstance class
