+++ modules/shadertools/src/lib/shader-module/shader-module.ts
@@ -4,7 +4,8 @@
 
 import {NumberArray} from '@math.gl/types';
 import {UniformFormat} from '../../types';
-import {PropType} from '../filters/prop-types';
+import {PropType, PropValidator} from '../filters/prop-types';
+
 
 export type UniformValue = number | boolean | Readonly<NumberArray>; // Float32Array> | Readonly<Int32Array> | Readonly<Uint32Array> | Readonly<number[]>;
 
@@ -53,6 +54,18 @@ export type ShaderModule<
 
   /** Internal */
   normalized?: boolean;
+  instantiated?: boolean;
+  propValidators?: Record<string, PropValidator>;
+  getModuleUniforms: Function;
+  // dependencies: ShaderModuleInstance[];
+  // deprecations: ShaderModuleDeprecation[];
+  // defines: Record<string, string | number>;
+  injections: {
+    vertex: Record<string, ShaderInjection>;
+    fragment: Record<string, ShaderInjection>;
+  };
+  // uniforms: Record<string, PropValidator> = {};
+  // uniformTypes: Record<string, PropValidator> = {};
 };
 
 /** Use to generate deprecations when shader module is used */
